Move concatenate bothmethod to base.AwkwardArray
